==> analysis/11_wc_astro_wgcna.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Load Libraries

```{r, warnings=F, message=F}
library(Seurat)
library(WGCNA)
library(cluster)
library(parallelDist)
library(ggsci)
library(emmeans)
library(lme4)
library(ggbeeswarm)
library(genefilter)
library(tidyverse)
library(reshape2)
library(igraph)
library(gProfileR)
library(ggpubr)
library(cowplot)
library(here)
library(tidygraph)
library(ggraph)
library(ggExtra)
```

```{r include = FALSE}
knitr::opts_chunk$set(message = FALSE, warnings = FALSE)
source(here("code/sc_functions.R"))
```

# Extract Cells for WGCNA
```{r, warnings=FALSE, message=FALSE, include=FALSE}
fgf.glia.wc <- readRDS(here("data/filtglia.RDS"))
fgf.glia.wc %>%
  subset(ident = c("Astro")) %>%
  SCTransform(return.only.var.genes = F, verbose = F, variable.features.n = 5000) -> ventric
rm(fgf.glia.wc)
```

# Calculate softpower
```{r}
enableWGCNAThreads()
datExpr <- as.matrix(t(ventric[["SCT"]]@scale.data[ventric[["SCT"]]@var.features, ]))
gsg <- goodSamplesGenes(datExpr, verbose = 3)
powers <- c(c(1:10), seq(from = 12, to = 40, by = 2))
sft <- pickSoftThreshold(datExpr,
  dataIsExpr = TRUE, powerVector = powers, corOptions = list(use = "p"),
  networkType = "signed"
)
cex1 <- 0.9
plot(sft$fitIndices[, 1], -sign(sft$fitIndices[, 3]) * sft$fitIndices[, 2], xlab = "Soft Threshold (power)", ylab = "Scale Free Topology Model Fit, signed R^2", type = "n", main = paste("Scale independence"))
text(sft$fitIndices[, 1], -sign(sft$fitIndices[, 3]) * sft$fitIndices[, 2], labels = powers, cex = cex1, col = "red")
abline(h = 0.80, col = "red")
# Mean Connectivity Plot
plot(sft$fitIndices[, 1], sft$fitIndices[, 5], xlab = "Soft Threshold (power)", ylab = "Mean Connectivity", type = "n", main = paste("Mean connectivity"))
text(sft$fitIndices[, 1], sft$fitIndices[, 5], labels = powers, cex = cex1, col = "red")
```

# Generate TOM
```{r}
softPower <- 14
SubGeneNames <- colnames(datExpr)
adj <- adjacency(datExpr, type = "signed", power = softPower)
diag(adj) <- 0
TOM <- TOMsimilarityFromExpr(datExpr, networkType = "signed", TOMType = "signed", power = softPower, maxPOutliers = 0.05)
colnames(TOM) <- rownames(TOM) <- SubGeneNames
dissTOM <- 1 - TOM
geneTree <- hclust(as.dist(dissTOM), method = "complete") # use complete for method rather than average (gives better results)
plot(geneTree, xlab = "", sub = "", cex = .5, main = "Gene clustering", hang = .001)
```

# Identify Modules
```{r Cut Gene Tree}
minModuleSize <- 15
x <- 2

dynamicMods <- cutreeDynamic(
  dendro = geneTree, distM = as.matrix(dissTOM),
  method = "hybrid", pamStage = F, deepSplit = x,
  minClusterSize = minModuleSize
)
dynamicColors <- labels2colors(dynamicMods) # label each module with a unique color
plotDendroAndColors(geneTree, dynamicColors, "Dynamic Tree Cut",
  dendroLabels = FALSE, hang = 0.03, addGuide = TRUE, guideHang = 0.05,
  main = "Gene dendrogram and module colors"
) # plot the modules with colors
```

# Calculate Eigengenes 
```{r merge close modules}
MEs <- moduleEigengenes(datExpr, dynamicColors)$eigengenes 
ME1 <- MEs
row.names(ME1) <- row.names(datExpr)
MEDiss <- 1 - cor(MEs)
METree <- hclust(as.dist(MEDiss), method = "average")
plot(METree, main = "Clustering of module eigengenes", xlab = "", sub = "")
MEDissThres <- 0.2
abline(h = MEDissThres, col = "red")
```

# Merge Modules
```{r}
merge <- mergeCloseModules(datExpr, dynamicColors, cutHeight = MEDissThres, verbose = 3)
mergedColors <- merge$colors
mergedMEs <- merge$newMEs
moduleColors <- mergedColors
MEs <- mergedMEs
modulekME <- signedKME(datExpr, MEs)
```

# Plot merged modules
```{r}
plotDendroAndColors(geneTree, cbind(dynamicColors, mergedColors),
  c("Dynamic Tree Cut", "Merged dynamic"),
  dendroLabels = FALSE, hang = 0.03,
  addGuide = TRUE, guideHang = 0.05
)
moduleColors <- mergedColors
MEs <- mergedMEs
modulekME <- signedKME(datExpr, MEs)
```

# Generate function to look up genes in each module
```{r}
modules <- MEs
c_modules <- data.frame(moduleColors)
row.names(c_modules) <- colnames(datExpr) 
module.list.set1 <- substring(colnames(modules), 3)
index.set1 <- 0
Network <- list()
for (i in 1:length(module.list.set1)) {
  index.set1 <- which(c_modules == module.list.set1[i])
  Network[[i]] <- row.names(c_modules)[index.set1]
}
names(Network) <- module.list.set1
```

# Filter metadata table and correlate with eigengenes
```{r, fig.height=6, fig.width=5}
nGenes <- ncol(datExpr)
nSamples <- nrow(datExpr)
MEs <- orderMEs(MEs)
MEs %>% select(-MEgrey) -> MEs
var <- model.matrix(~ 0 + ventric$trt)
moduleTraitCor <- cor(MEs, var, use = "p")
cor <- moduleTraitCor[abs(moduleTraitCor[, 1]) > .2, ]
moduleTraitPvalue <- corPvalueStudent(moduleTraitCor, nSamples)
cor <- melt(cor)
ggplot(cor, aes(Var2, Var1)) + geom_tile(aes(fill = value), colour = "white") + 
  scale_fill_gradient2( midpoint = 0, low = "blue", mid = "white",
  high = "red", space = "Lab", name = "Correlation \nStrength") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) + xlab("Treatment") + ylab(NULL)
```

# Run linear regression to calculate sig diff mods
```{r, warnings=F, message=F}
data <- data.frame(MEs,
  trt = ventric$trt,
  sample = as.factor(ventric$sample),
  batch = as.factor(ventric$batch)
)

mod <- lapply(colnames(MEs)[grepl("^ME", colnames(MEs))], function(me) {
  tryCatch({
    mod <- lmer(MEs[[me]] ~ trt + (1 | batch) + (1 | sample), data = data)
    pairwise <- emmeans(mod, pairwise ~ trt)
    plot <- data.frame(plot(pairwise, plotIt = F)$data)
    sig <- as.data.frame(pairwise$contrasts)
    return(sig)
  }, error = function(err) {
    print(err)
  })
})

names(mod) <- colnames(MEs)[grepl("^ME", colnames(MEs))]
sig <- bind_rows(mod, .id="id")
write_csv(sig, path = here("output/glia/wgcna/wc_astro_linmod.csv"))


mod <- data.frame(unlist(mod))
mod %>%
  add_rownames("test") %>%
  separate(test, c("mod", "measure")) %>%
  dcast(measure ~ mod, value = unlist.mod.) %>%
  as.data.frame() %>%
  t() -> test
colnames(test) <- test[1, ]
data.frame(test) %>%
  add_rownames("mod") %>%
  slice(2:nrow(.)) %>%
  select(p, estimate, mod) %>%
  mutate(p = as.numeric(as.character(p)), estimate = as.numeric(as.character(estimate))) %>%
  arrange(log10(p)*abs(estimate)) %>%
  mutate(pubname = paste0("sc-astro-M", paste0(seq_len(length(estimate)))))  -> astro_mods

astro_mods %>% filter(p < 0.05, abs(estimate) > 0.005) %>% pull(pubname) -> sigmods

rename_mods <- astro_mods$mod
names(rename_mods) <- astro_mods$pubname
MEs <- rename(MEs, !!!rename_mods)
data <- rename(data, !!!rename_mods)
write_csv(data.frame(pubname = names(rename_mods), color = as.character(rename_mods)),
          path = here("output/glia/wgcna/wc_astro_translatelabels.csv"))

```

# hubgene information
```{r}
hubgenes <- lapply(seq_len(length(Network)-1), function(x) {
  dat <- modulekME[Network[[x]], ]
  dat <- dat[order(-dat[paste0("kME", names(Network)[x])]), ]
  gene <- data.frame(gene=rownames(dat),kme=dat[,x]) 
  return(gene)
})

names(hubgenes) <-  names(rename_mods)[match(names(Network)[-length(Network)], 
                                             gsub(as.character(rename_mods),pattern = "ME",replacement = ""))]
d <- dplyr::bind_rows(hubgenes, .id = "id")
write_csv(d, path = here("output/glia/wgcna/wc_astro_wgcna_genemods.csv"))
```

# boxplot summary of diff exp modules
```{r, fig.height=2, fig.width=4}
data <- melt(data, id.vars = c("trt", "sample"))
data %>% filter(variable %in% sigmods) -> data
data$trt <- gsub(data$trt, pattern = "PF", replacement = "Veh")
boxplot <- ggplot(data = data, 
                  aes(x = fct_relevel(variable,"sc-astro-M1","sc-astro-M3",
                                      "sc-astro-M2","sc-astro-M9","sc-astro-M5","sc-astro-M6"), 
                      y = as.numeric(value))) +
  geom_boxplot(aes(fill=trt), width=0.5, notch=T, outlier.shape = NA) +
  theme_pubr(legend="right") + 
  geom_hline(yintercept = 0, linetype = "dashed") + xlab(NULL) +
  ylab("ME Expression") + 
  theme(axis.text.x = element_blank(), 
        axis.title.y = element_text(size=10, face="bold"),
        legend.text = element_text(size=8, face="bold"),
        legend.title = element_text(size=10, face="bold"), 
        plot.margin = unit(c(0, 0, 0, 0), "cm")) +
  ylim(c(-0.025,0.025)) + scale_fill_manual(name="Treatment", values = c("gray30", "gray80")) + theme_figure

boxplot
```

# go term enrichment
```{r}
goterms <- lapply(hubgenes[sigmods], function(x) {
  x <- gprofiler(x,
    ordered_query = T, organism = "mmusculus", significant = T, custom_bg = colnames(datExpr),
    src_filter = c("GO:BP", "REAC", "KEGG"), hier_filtering = "strong",
    min_isect_size = 2,
    sort_by_structure = T, exclude_iea = T,
    min_set_size = 10, max_set_size = 300, correction_method = "fdr"
  )
  x <- x[order(x$p.value), ]
  return(x)
})

godat <- bind_rows(goterms, .id="id")
write_csv(godat, path = here("output/glia/wgcna/wc_astro_wgcna_goterms.csv"))
```

# test overlap with nuclei modules
```{r, fig.width=4, fig.height=4}
nuc_mods <- read_csv(file = here("output/glia/wgcna/astro_wgcna_genemodules.csv"))
nuc_mods %>%
  as.data.frame() %>%
  filter(id %in% c(paste0("sn-astro-M", seq_len(4)))) %>%
  dplyr::group_by(id) %>%
  dplyr::group_split() %>%
  map("gene") -> nuc_gene

nuc_mods %>% as.data.frame() %>%
  filter(id %in% c(paste0("sn-astro-M", seq_len(4)))) %>%
  dplyr::group_by(id) %>% group_keys() %>% pull() -> names(nuc_gene)

wc_nuc_overlap <- sapply(nuc_gene, function(x) {
  sapply(hubgenes[sigmods], function(y) {
    1 - phyper(sum(x %in% y$gene), length(y$gene), 5000 - length(y$gene), length(x), log.p = F)
  })
})


wc_nuc_overlap <- reshape2::melt(wc_nuc_overlap)
wc_nuc_overlap %>%
  mutate(value = p.adjust(wc_nuc_overlap$value, n = dim(wc_nuc_overlap)[1] * dim(wc_nuc_overlap)[2])) %>%
  mutate(sig = if_else(value > 0.05, "",
    if_else(.05 > value & value > .01, "*",
      if_else(.01 > value & value > .001, "**",
        if_else(.001 > value, "***", "")
      )
    )
  )) -> wc_nuc_overlap

write_csv(wc_nuc_overlap, path = here("output/glia/wgcna/astro_mod_overlaps_pval.csv"))

overlap <- ggplot(wc_nuc_overlap, aes(x = fct_relevel(Var1, "sc-astro-M1", "sc-astro-M3","sc-astro-M2","sc-astro-M9",
                                                      "sc-astro-M5","sc-astro-M6"),
                                      y = fct_relevel(Var2, "sn-astro-M4",
                                                      "sn-astro-M3","sn-astro-M2","sn-astro-M1"))) +
  geom_tile(size = 1, color = "white", fill="grey99") +
  geom_point(aes(size = if_else(-log10(value)<1.3,true = 0, false = -log10(value))), alpha=0.25) + 
  scale_size(name= expression(bold(-log[10] ~ pvalue)), breaks = c(2,4,8)) + 
  scale_fill_manual(values=c("black","red"), guide=F) + 
  theme_pubr(legend = "right") + xlab(NULL) + ylab(NULL) + 
  theme(axis.text.x = element_text(angle=45, hjust=1), 
        legend.title = element_text(size=8), 
        legend.text = element_text(size=7, face="bold"),
        plot.margin = unit(c(0, 0, 0, 0), "cm")) + theme_figure

overlap 
```

# plot as grid
```{r, fig.width=5, fig.height=5}
plot_grid(boxplot, overlap, align="v", axis="tblr", labels=c("f"), ncol=1)
```

# Plot gene networks
```{r, fig.width=5, fig.height=5}
color <- c("sc-astro-M1","sc-astro-M3")
lapply(color, function(col) {
  maxsize <- 15
  hubs <- data.frame(genes=hubgenes[[col]]$gene[1:maxsize], kme = hubgenes[[col]]$kme[1:maxsize], mod = rep(col,15))
}) -> hub_plot 


hub_plot <- lapply(hub_plot, function(x) {
  adj[as.character(x$genes), as.character(x$genes)] %>% 
  graph.adjacency(mode = "undirected", weighted = T, diag = FALSE) %>%
  as_tbl_graph(g1) %>% upgrade_graph() %>% activate(nodes) %>% dplyr::mutate(mod=x$mod) %>%
  dplyr::mutate(kme=x$kme)  %>% 
    dplyr::mutate(color = ifelse(name %in% c("Slc1a2","Slc1a3","Slc6a11","Aqp4","Sparcl1","Sparc","Psap","Clu","Plat","Cpe"), yes="red",no="black")) %>%
  activate(edges)}
)

set.seed("139")
plotlist <- lapply(hub_plot, function(x) {
  print(ggraph(x,  layout = 'fr') + 
  geom_edge_link(color="darkgrey", aes(alpha = weight), show.legend = F)  + 
  scale_edge_width(range = c(0.2, .5)) + geom_node_text(aes(label = name, color=color), fontface="bold", size=3, repel=T) +
  scale_color_manual(values=c("gray0","red")) + guides(color=FALSE) +
  geom_node_point(shape=21, alpha=0.5, fill="grey70", size=4) + 
  theme_graph() + theme(legend.position = "none", plot.title = element_text(hjust=0.5, vjust=1)) + 
    coord_cartesian(clip = "off"))
})
```

# arrange figure
```{r, fig.height=6, fig.width=13}
genenet <- plot_grid(plotlist[[1]], plotlist[[2]], ncol=1, labels = c("f"), scale=0.8)
nets <- plot_grid(boxplot, overlap, align="v", axis="tblr", labels=c("e"), ncol=1, scale=0.95, rel_heights = c(2,1))
wc_astro_res <- plot_grid(nets, genenet, rel_widths = c(1.5,1), nrow=1)

load(here("data/figures/fig4/fig4_nuc_top.RData"))
fig4 <- plot_grid(top_astro, wc_astro_res, ncol=1, align="hv", rel_heights = c(1,1.25), axis="lr")
transcriptomics_4 <- plot_grid(top_astro, fig4, rel_widths = c(1,1))
transcriptomics_4
```

# Aqp4/Gfap Quantification
```{r}
readxl::read_xlsx(here("data/mouse_data/fig4/191118_Aqp4_Gfap.xlsx"), range="A5:E9") %>%
  select(1,2,4,5) %>% 
  reshape2::melt() %>% 
  mutate(gene = c(rep("Aqp4", 8), rep("Gfap",8))) %>% 
  separate(variable, "[...]", into = "trt") -> aq_gf_quants
aq_gf_quants %>% dplyr::group_by(trt,  gene) %>% dplyr::summarise(mean = mean(value), sd = sd(value), se = sd/sqrt(length(value))) %>%
  ggplot(aes(x=gene, y=mean, fill=fct_relevel(trt,"Veh","FGF1"), color = fct_relevel(trt,"Veh","FGF1"))) + 
  geom_col(width=0.9, alpha=0.75, colour="black", position="dodge") +
  geom_errorbar(aes(x=gene, ymin = mean-se, ymax=mean+se), width=0.2, position = position_dodge(0.9)) +
  geom_jitter(data = aq_gf_quants, inherit.aes = F, aes(x=gene, y=value, fill=fct_relevel(trt,"Veh","FGF1")), 
              alpha=0.85, shape=21, position = position_jitterdodge(0.2)) + xlab(NULL) + 
  ylab(expression(bold(Volume~(mu*m^3)))) + scale_fill_manual("Treatment", values=c("gray80","gray30")) + 
  scale_color_manual("Treatment", values=c("gray80","gray30")) +
  scale_y_continuous(labels = function(x) format(x, scientific = TRUE)) +
  geom_signif(y_position=c(1.5e6), xmin=c(0.9,1.9), xmax=c(1.1,2.1),
              annotation=c("*","**"), tip_length=0, size = 0.5, textsize = 5, color="black") +
  theme_pubr() + theme(legend.position="none", axis.text.x=element_text(angle=45, hjust=1)) + theme_figure -> aq_gf
aqp4_gfap <- cowplot::plot_grid(ggplot() + theme_void(), aq_gf, nrow=1, scale=c(1,0.75), labels="g", rel_widths = c(2.05,1))
aqp4_gfap
```

# Agrp/Gfap overlap
```{r}
readxl::read_xlsx(here("data/mouse_data/fig4/191118_GFAP_AGRP.xlsx"), range="A5:B9") %>%
  reshape2::melt() -> ga_quants
ga_quants %>% dplyr::group_by(variable) %>% dplyr::summarise(mean = mean(value), sd = sd(value), se = sd/sqrt(length(value))) %>%
  ggplot(aes(x = variable, y = mean, fill = variable, color = variable)) + 
  geom_col(width=1, alpha=0.75, colour="black", position="dodge") +
  geom_errorbar(aes(x=variable, ymin = mean-se, ymax=mean+se), width=0.2, position = position_dodge(0.9)) + 
  geom_jitter(data = ga_quants, inherit.aes = F, aes(x=variable, y=value, fill=variable), 
              alpha=0.5, shape=21, position = position_jitterdodge(.5)) + xlab(NULL) + 
  geom_signif(y_position=c(1000), xmin=c(1.2), xmax=c(1.8),
              annotation=c("**"), tip_length=0, size = 0.5, textsize = 5, color="black") +
  ylab(expression(bold(GFAP/Agrp~contact~volume~(mu*m^3/hpf)))) + scale_fill_manual(values=c("gray80","gray30"), name="") + 
  scale_color_manual(values=c("gray80","gray30"), name="") +
  theme_pubr() + theme(legend.position = "none", axis.text.x = element_text(angle=45,hjust=1)) + theme_figure -> gfap_agrp
gfap_agrp <- cowplot::plot_grid(ggplot() + theme_void(), gfap_agrp, nrow=1, scale=c(1,0.75), labels="h", rel_widths = c(1.5,1))
gfap_agrp
```

# EM of PAPs quantification
```{r}
readxl::read_xlsx(here("data/mouse_data/fig4/EM_glial_contact.xlsx"), range="A4:B8") %>%
  reshape2::melt() -> em_quants
em_quants %>% dplyr::group_by(variable) %>% dplyr::summarise(mean = mean(value), sd = sd(value), se = sd/sqrt(length(value))) %>%
  ggplot(aes(x=variable, y=mean, fill=variable, color = variable)) +
  geom_col(width=1, alpha=0.75, colour="black", position="dodge") +
  geom_errorbar(aes(x=variable, ymin = mean-se, ymax=mean+se), width=0.2, position = position_dodge(0.9)) +
  geom_jitter(data = em_quants, inherit.aes = F, aes(x=variable, y=value, fill=variable), 
              alpha=0.5, shape=21, position = position_jitterdodge(.5)) + xlab(NULL) + 
  geom_signif(y_position=c(85), xmin=c(1.2), xmax=c(1.8),
              annotation=c("***"), tip_length=0, size = 0.5, textsize = 5, color="black") +
  ylab("% Astrocytic coverage") + scale_fill_manual("Treatment", values=c("gray80","gray30")) + 
  scale_color_manual("Treatment", values=c("gray80","gray30")) +
  theme_pubr() + theme(legend.position = "none",axis.text.x = element_text(angle=45, hjust=1)) + 
  coord_cartesian(clip="off") + theme_figure -> em_bp
pap <- cowplot::plot_grid(ggplot() + theme_void(), em_bp, nrow=1, scale=c(1,0.75), labels="i", rel_widths = c(1.5,1))
pap
```

# Generate full figure
```{r, fig.width=13, fig.height=10}
validation <- plot_grid(aqp4_gfap, gfap_agrp, pap, nrow=1, align="h", axis="tb", rel_widths = c(1.5,1,1))
plot_grid(transcriptomics_4, validation, ncol=1, align="h", axis="tblr", rel_heights = c(1.75,1))
ggsave(filename = here("data/figures/fig4/fig4.tiff"), width = 13, h=10, compression="lzw")
```
