==> analysis/13_olig_pseudotime.Rmd <==
---
title: "Olig Pseudotime"
output: html_notebook
---

```{r}
library(here)
library(Seurat)
library(monocle)
library(ggplot2)
library(tidyverse)
library(rstatix)
library(ggpubr)
library(ggsci)
library(ggrepel)
library(reshape2)
library(cowplot)
library(ggpubr)
```

```{r include = FALSE}
knitr::opts_chunk$set(message = FALSE, warnings = FALSE)
source(here("code/sc_functions.R"))
```

```{r include=FALSE}
addSmallLegend <- function(myPlot, barwidth = 5, textSize = 8, spaceLegend = 0.1, title.position = "top", barheight = .5) {
    myPlot +
        guides(color = guide_colorbar(barwidth = barwidth, title.position = title.position, barheight = barheight)) +
        theme(legend.title = element_text(size = textSize), 
              legend.text  = element_text(size = textSize),
              legend.key.size = unit(spaceLegend, "lines"))
}
```

```{r, fig.width=10, fig.height=3}
olig <- readRDS(here("data/glia/olig_labeled.RDS"))
olig_plot <- as.data.frame(Embeddings(olig, reduction = "umap"))
olig_plot$trt <- olig$trt
olig_plot$type <- Idents(olig)
label.df <- data.frame(cluster=levels(olig_plot$type),label=levels(olig_plot$type))
label.df_2 <- olig_plot %>% 
  dplyr::group_by(type) %>% 
  dplyr::summarize(x = median(UMAP_1), y = median(UMAP_2))

a <- ggplot(olig_plot, aes(UMAP_1, UMAP_2, colour = trt)) + 
  geom_point(alpha = 0.5, size=.5) + scale_color_manual(values=c("#000000","#999999"), name="") +  
  guides(colour = guide_legend(override.aes = list(size=2))) + theme_pubr() + theme(legend.position = c(0.3, 0.25), legend.background=element_blank())
b <- ggplot(olig_plot, aes(UMAP_1, UMAP_2, colour = type)) + 
  geom_point(alpha = 0.5, size=.5) + scale_colour_discrete(name="Treatment") +
  geom_label_repel(data = label.df_2, aes(label = type, x=x, y=y), size=3, fontface="bold", inherit.aes = F) +
  guides(colour = guide_legend(override.aes = list(size=5))) + theme_pubr() + theme(legend.position = "none") 
plot_grid(a,b)
```

# Test difference in cell numbers
```{r}
cell<-as.data.frame.matrix(table(olig$orig.ident, olig@active.ident))
cell$trt<-as.factor(sapply(strsplit(rownames(cell),"_"),"[",2))
cell<-melt(cell)
stat.test <- cell %>%
  group_by(variable) %>%
  t_test(value ~ trt) %>%
  adjust_pvalue() %>%
  add_significance("p.adj")

cell %>% dplyr::group_by(trt, variable) %>% 
  dplyr::summarise(mean=mean(value), sd = sd(value), se = sd/sqrt(length(value))) %>% 
  mutate(signif = stat.test$p.adj.signif) %>% 
  mutate(signif = ifelse(trt == "FGF1", yes = NA, no = signif)) %>% ungroup() -> plotval
write.csv(plotval, file="olig_ttest_padj.csv")

plotval %>% mutate(variable = fct_relevel(variable, c("OPC","COP", "NFOL","MFOL","MOL1"))) %>% 
  mutate(trt = fct_relevel(trt, c("Vehicle","FGF1")))-> plotval 

ggplot(plotval, aes(x = variable, y = mean, fill = trt)) + 
    geom_bar(position=position_dodge(), stat="identity") + 
    geom_errorbar(aes(ymin=mean-se, ymax=mean+se),size=.3,width=.2,position=position_dodge(.9)) +
    xlab(NULL) + scale_fill_manual(values=c("gray80","gray30")) +
    ylab("Number of cells") +
    ggpubr::theme_pubr(legend = "none") +
    theme(axis.text.x = element_text(angle=45, hjust=1)) + 
    geom_signif(y_position=c(plotval %>% dplyr::group_by(variable) %>% dplyr::summarise(max = max(mean)) %>% pull(max) + 50), 
                xmin = c(seq(0.9,4.9, by = 1)), xmax=c(seq(1.1,5.1, by = 1)),
                annotation=c(plotval %>% slice(6:10) %>% pull(signif) %>% as.character() %>% toupper())[c(1,4,5,3,2)], 
                tip_length=0, size = 0.5, textsize = 3, color="black", vjust = -1) + coord_cartesian(clip="off") -> oligttest
oligttest
```

# Prep data for pseudotime analysis
```{r}
cds <- as.CellDataSet(olig)
cds <- estimateSizeFactors(cds)
cds <- estimateDispersions(cds)
cds <- detectGenes(cds, min_expr = 0.1)
fData(cds)$use_for_ordering <-
    fData(cds)$num_cells_expressed > 0.1 * ncol(cds)

cds <- reduceDimension(cds,
                              max_components = 2,
                              norm_method = 'log',
                              num_dim = 2,
                              reduction_method = 'tSNE',
                              verbose = T)
cds <- clusterCells(cds, verbose = T)
cds <- clusterCells(cds,
                 rho_threshold = 150,
                 delta_threshold = 15,
                 skip_rho_sigma = T,
                 verbose = F)
plot_cell_clusters(cds, label_groups_by_cluster=FALSE,  color_cells_by = "Cluster")
```

# Run monocle
```{r}
olig_expressed_genes <-  row.names(subset(fData(cds), num_cells_expressed >= 10))

clustering_DEG_genes <-
    differentialGeneTest(cds[olig_expressed_genes,],
          fullModelFormulaStr = '~predicted.id',
          cores = 10)

olig_ordering_genes <-
    row.names(clustering_DEG_genes)[order(clustering_DEG_genes$qval)][1:500]

cds <-
    setOrderingFilter(cds,
        ordering_genes = olig_ordering_genes)

cds <-
    reduceDimension(cds, method = 'DDRTree')

cds <-
    orderCells(cds)

cds <-
    orderCells(cds, root_state = 2)

olig$pseudo <- cds$Pseudotime
plot_cell_trajectory(cds,color_by = "predicted.id")
ggsave(here("data/figures/supp/monocle_by_celltype.pdf"))
plot_cell_trajectory(cds, markers = "Gpr17")
ggsave(here("data/figures/supp/monocle_by_gpr17.pdf"))
plot_cell_trajectory(cds, color_by = "Pseudotime")
ggsave(here("data/figures/supp/monocle_by_pseudo.pdf"))
plot_cell_trajectory(cds, color_by = "trt", alpha=0.5)
ggsave(here("data/figures/supp/monocle_by_trt.pdf"))
```

# Plot pseudotime on umap plots
```{r, fig.height=5, fig.width=7}
olig_plot$pseudo <- olig$pseudo
olig_plot$gpr17 <- as.numeric(olig@assays[["SCT"]]@data["Gpr17",])

ggplot(olig_plot, aes(UMAP_1, UMAP_2, colour = pseudo)) + 
  geom_point(alpha = 0.5, size=.5) + ggsci::scale_color_material(name="Pseudotime", guide = guide_colorbar(title.position = "top"), palette = "blue-grey") +
  ggpubr::theme_pubr() + xlab(NULL) + ylab(NULL) +
  theme(legend.position = c(0.3,0.25), legend.direction =  "horizontal",legend.title = element_text(hjust=0.5), legend.background = element_blank()) -> olig_pseudo

ggplot(olig_plot, aes(UMAP_1, UMAP_2, color = gpr17)) + 
  geom_point(alpha = 0.75, size=.5) + ggsci::scale_color_material(name="Gpr17 Expression", palette = "deep-orange",
                                                                 guide = guide_colorbar(title.position = "top")) +
  ggpubr::theme_pubr() + xlab(NULL) + ylab(NULL) +
  theme(legend.position = c(0.3,0.25), legend.direction =  "horizontal",legend.title = element_text(hjust=0.5), legend.background = element_blank()) -> olig_gpr17

sc_olig <- cowplot::plot_grid(b, addSmallLegend(olig_pseudo), oligttest, addSmallLegend(olig_gpr17), ncol=2, labels="auto", scale=0.9, align="hv")
sc_olig
```

```{r, fig.width=12, fig.height=8}
readxl::read_xlsx(path = here("data/mouse_data/fig6/191118_Gpr17.xlsx"), range="A4:B10") %>%
  reshape2::melt() %>% na.omit() -> gpr17

gpr17 %>% dplyr::group_by(variable) %>% dplyr::summarise(mean=mean(value), sd = sd(value), se = sd/sqrt(length(value))) %>%
  ggplot(aes(x=variable, y=mean, fill=variable, color=variable)) + 
  geom_col(width=1, alpha=0.75, colour="black", position="dodge") +
  geom_errorbar(aes(x=variable, ymin = mean-se, ymax=mean+se), width=0.2, position=position_dodge(.9), size=1) +
  geom_jitter(data = gpr17, inherit.aes = F, aes(x=variable, y=value, fill=variable), 
              alpha=0.5, shape=21, position = position_jitterdodge(.5)) + xlab(NULL) +
  #geom_text(position = position_dodge2(width=.9, preserve="single"), aes(y=value+se+1), face = "bold", size=8) + 
  ylab("GPR17+ Cells Per mm²") + scale_fill_manual("Treatment", values=c("gray80","gray30")) + 
  scale_color_manual("Treatment", values=c("gray80","gray30"))+ theme_classic() +
  theme(legend.position = "none", legend.background = element_blank(), axis.text.x = element_blank(), axis.ticks.x = element_blank()) + 
  geom_signif(y_position=max(gpr17$value), xmin=1.2, xmax=1.8,
              annotation=c("*"), tip_length=0, size = 0.5, textsize = 6, color="black") + coord_cartesian(clip="off") -> gpr17_bp

olig_val <- cowplot::plot_grid(ggplot() + theme_void(), gpr17_bp, nrow=1, scale=0.9, labels=c("e","f"), rel_widths = c(2,1))
cowplot::plot_grid(sc_olig, olig_val, ncol=1, align="hv", rel_heights = c(1.75,1))
#ggsave("figure_6.tiff", width=12, h=8, dpi=600, compression = "lzw")
```

```{r echo=FALSE, include=FALSE}
detach("package:here", unload = T)
library(here)
save.image(file = here("data/glia/olig_alldata.RData"))
```
